refactor: move test util dottedLookup to utils file
